Modified home page init.js file to point to surveys URL
